requirements.txt
# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-cov==4.1.0

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
faker==20.1.0
pydantic==2.5.3

# Reporting
allure-pytest==2.13.2

# Data validation
jsonschema==4.20.0

# Performance Testing
locust==2.17.0
